+++ 07-http_gordle/5_play/internal/session/game.go
@@ -1,22 +1,29 @@
 package session
 
 import (
-	"fmt"
-
 	"learngo-pockets/httpgordle/internal/gordle"
 )
 
+// A GameID represents the ID of a game
+type GameID string
+
 // Game contains the information about a game.
 type Game struct {
-	ID           GameID
-	Gordle       gordle.Game
-	AttemptsLeft byte
-	Guesses      []Guess
-	Status       Status
-}
+	// ID is the identified of a game.
+	ID GameID
 
-// A GameID represents the ID of a game.
-type GameID string
+	// The game of Gordle that is being played.
+	Gordle gordle.Game
+
+	// AttemptsLeft counts the number of attempts left before the game is over.
+	AttemptsLeft byte
+
+	// Guesses is the list of past guesses, and their feedback.
+	Guesses []Guess
+
+	// Status tells whether the game is playable.
+	Status Status
+}
 
 // Status is the current status of the game and tells what operations can be made on it.
 type Status string
@@ -32,6 +39,3 @@ type Guess struct {
 	Word     string
 	Feedback string
 }
-
-// ErrGameOver can signal when the game cannot be played any further because it is over.
-var ErrGameOver = fmt.Errorf("game over")
